requirements.txt
fastapi
mangum
playwright
pydantic
httpx
pandas
numpy
PyPDF2
pdfplumber
openpyxl
pillow
matplotlib
seaborn
plotly
scikit-learn
python-dotenv
aiofiles
beautifulsoup4
lxml
requests

